# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Twitter like application",
    "version": "1.0.0",
    "title": "Swagger Petstore"
  },
  "host": "software-architecture-22.herokuapp.com",
  "basePath": "/api",
  "tags": [
    {
      "name": "users",
      "description": "CRUD operations about users"
    },
    {
      "name": "auth",
      "description": "Login endpoint"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new user to the database",
        "description": "",
        "operationId": "addUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be added to the database",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string", "format": "email" },
                "username": { "type": "string" },
                "password": { "type":  "string", "format": "password"},
                "name": { "type": "string" },
                "surname": { "type": "string" },
                "bio": { "type": "string" }
              },
              "required": [
                "email",
                "username",
                "name",
                "surname"
              ]
            }
          }
        ],
        "responses": {
          "200": { "description": "Ok"},
          "409": { "description": "Datos únicos formulario ya en uso" },
          "500": { "description": "Internal server error"}
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user by user name",
        "description": "",
        "operationId": "getUserByUsername",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be fetched.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "successful operation", "schema": {"$ref": "#/definitions/User"}},
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error"}
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": { "description": "Successful operation", "schema": {"$ref": "#/definitions/User"}},
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error"}
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Successful operation"},
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error"}
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User authentication",
        "tags": [
          "auth"
        ],
        "description": "",
        "operationId": "loginUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": { "description": "Successful operation", "schema": {"$ref": "#/definitions/JWT"}},
          "401": { "description": "Unauthorized request"},
          "404": { "description": "Email not found" },
          "500": { "description": "Internal server error"}
        }
      }
    },
    "/user/{username}/follow": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add user to follow",
        "description": "This can only be done by the logged in user.",
        "operationId": "followUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "usernameFollowed",
            "in": "formData",
            "description": "username to follow",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Successful operation", "schema": {"$ref": "#/definitions/User"}},
          "404": { "description": "User not found" },
          "500": { "description": "Internal server error"}
        }
      }
    }

  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "email": { "type": "string", "format": "email" },
        "password": { "type": "string" },
        "username": { "type": "string" },
        "name": { "type": "string" },
        "surname": { "type": "string" },
        "bio": { "type": "string" },
        "follows": { "type": "array", "items": { "type": "string"}},
        "followers": { "type": "array", "items": { "type": "string"}}
      },
      "required": [
        "email",
        "password",
        "username",
        "name",
        "surname"
      ]
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": { "type": "string", "format": "email" },
        "password": { "type": "string" }
      }
    },
    "JWT": {
      "type": "object",
      "properties": {
        "token": { "type": "string" }
      }
    }
  }
}